+++ include/bmp/bmp.h
@@ -6,6 +6,7 @@
 
 int bmp_write(const char *filename, int *fb, int width, int height);
 
-int bmp_load(const char *filename, int **buffer, long *width, long *height);
+int bmp_load(const char *filename, int **buffer, unsigned *width,
+	     unsigned *height);
 
 #endif
